+++ src/components/entries/entryTotalAmount.js
