+++ src/components/inputForm/inputFormItems/amount.js
