+++ src/components/totalAmount/totalAmount-util.js
