+++ src/components/totalAmount/totalAmount.js
