+++ src/js/core.js
@@ -1,7 +1,7 @@
-import { sharedState } from "../../../store/state.js";
+import { sharedState } from "../store/state.js";
 import { updateTotalAmounts } from "../components/totalAmount/totalAmount-util.js";
-import { renderOneEntiry } from "../../components/entries/entry-util.js";
-import { loadEntriesFromServer } from "../api.js";
+import { renderOneEntiry } from "../../src/components/entries/entry-util.js";
+import { loadEntriesFromServer } from "../../api.js";
 import { renderCategoryOptions } from "../components/inputForm/inputFormItems/categoryRender.js";
 
 const entries = sharedState.entries;
