+++ src/pages/main/stats/statsView.js
