+++ src/pages/main/viewSwitcher.js
