+++ src/setting/config.js
